base library should not have Core prefix, alpscore based tests are moved to alpsocre subdir
