+++ CMakeLists.txt
@@ -131,21 +131,23 @@ if(Testing)
     enable_testing()
 
     # Core-only tests: build and run regardless of EDLIB_WITH_ALPSCORE.
-    add_subdirectory(test/core)
-    add_test(HubbardCoreTest    test/core/HubbardCoreTest)
-    add_test(LanczosCoreTest    test/core/LanczosCoreTest)
-    add_test(SzSymmetryCoreTest test/core/SzSymmetryCoreTest)
-    add_test(NSymmetryCoreTest  test/core/NSymmetryCoreTest)
-    add_test(MpiCoreTest        test/core/MpiCoreTest)
+    add_subdirectory(test)
+    add_test(HubbardTest    test/HubbardTest)
+    add_test(LanczosTest    test/LanczosTest)
+    add_test(SzSymmetryTest test/SzSymmetryTest)
+    add_test(NSymmetryTest  test/NSymmetryTest)
+    if(USE_MPI)
+        add_test(MpiTest        test/MpiTest)
+    endif()
 
     # Legacy alps-using tests only when the option is enabled.
     if(EDLIB_WITH_ALPSCORE)
-        add_subdirectory(test)
-        add_test(SzSymmetryTest        test/SzSymmetryTest)
-        add_test(NSymmetryTest         test/NSymmetryTest)
-        add_test(HubbardModelTest      test/HubbardModelTest)
-        add_test(LanczosTest           test/LanczosTest)
-        add_test(StaticObservablesTest test/StaticObservablesTest)
+        add_subdirectory(test/alpscore)
+        add_test(SzSymmetryAlpsCoreTest        test/alpscore/SzSymmetryAlpsCoreTest)
+        add_test(NSymmetryAlpsCoreTest         test/alpscore/NSymmetryAlpsCoreTest)
+        add_test(HubbardModelAlpsCoreTest      test/alpscore/HubbardModelAlpsCoreTest)
+        add_test(LanczosAlpsCoreTest           test/alpscore/LanczosAlpsCoreTest)
+        add_test(StaticObservablesAlpsCoreTest test/alpscore/StaticObservablesAlpsCoreTest)
     endif()
 endif()
