+++ test/NSymmetry_Test.cpp
@@ -1,17 +1,14 @@
-//
-// Created by iskakoff on 21/08/16.
-//
+// Core-only NSymmetry parity test.
 
-#include "gtest/gtest.h"
+#include <edlib/NSymmetry.h>
+#include <edlib/Parameters.h>
 
-#include "edlib/alpscore/NSymmetry.h"
-#include "edlib/alpscore/EDParams.h"
+#include <gtest/gtest.h>
 
-TEST(NSymmetryTest, Initialization) {
-  alps::params p;
-  EDLib::define_parameters(p);
-  EDLib::Symmetry::NSymmetry sym(p);
-  while(sym.next_sector()) {
+TEST(NSymmetryCore, IndexMatchesIteration) {
+  edlib::Parameters p;
+  edlib::NSymmetry sym(p);
+  while (sym.next_sector()) {
     sym.init();
     int i = 0;
     while (sym.next_state()) {
